BoxCollider uses floats

+++ source/components/inc/Gng2D/components/collider.hpp
@@ -10,8 +10,8 @@ struct BoxCollider
     component_property_name(BoxCollider);
     component_property_is_detail;
 
-    int width{};
-    int height{};
+    float width{};
+    float height{};
 };
 } // namespace detail
